fix: refactor verification
