+++ src/types/verification/IRenderPersonal/index.ts
@@ -1,0 +1,6 @@
+import { IOptionInitialiceEntry } from "@ptypes/positions/assisted/IOptionInitialiceEntry";
+interface IRenderPersonal {
+  values: IOptionInitialiceEntry[];
+  isMobile: boolean;
+}
+export type { IRenderPersonal };
