+++ src/types/verification/IRenderPersonalTwo/index.ts
@@ -1,0 +1,6 @@
+import { IGeneralInformationEntry } from "@ptypes/positions/assisted/IGeneralInformationEntry";
+interface IRenderPersonalTwo {
+  values: IGeneralInformationEntry;
+  isMobile: boolean;
+}
+export type { IRenderPersonalTwo };
