+++ src/types/verification/IVerificationBoxes/index.ts
@@ -1,0 +1,7 @@
+import { IFormAddPosition } from "@ptypes/forms/verificationForm/IFormAddPosition";
+interface IVerificationBoxes {
+  updatedData: IFormAddPosition;
+  stepKey: number;
+}
+
+export type { IVerificationBoxes };
